fix: 좌표 값을 정밀하게 저장하기 위해 float → double로 타입 변경
좌표는 15번째 소수점까지 존재하는데, float 타입에 저장하면 자동으로 7번째 소수점까지 자른다.

+++ BE/src/timelines/entities/timeline.entity.ts
@@ -24,10 +24,10 @@ export class Timeline {
   @Column({ length: 255, nullable: true })
   image: string;
 
-  @Column({ name: 'coord_x', type: 'float', nullable: true })
+  @Column({ name: 'coord_x', type: 'double', nullable: true })
   coordX: number;
 
-  @Column({ name: 'coord_y', type: 'float', nullable: true })
+  @Column({ name: 'coord_y', type: 'double', nullable: true })
   coordY: number;
 
   @Column({ type: 'date' })
